add: calculate best level configuration (grids count to get letter biggest size on screen)

+++ Assets/SOW/Scripts/Core/UI/Screen/Game/Views/LetterView/LetterView.cs
@@ -28,7 +28,7 @@ namespace SoW.Scripts.Core.UI.Screen.Game.Views
         }
 
         [FoldoutGroup("API"), Button]
-        public void SetSize(int pixelSize)
+        public void SetSize(float pixelSize)
         {
             RT.sizeDelta = Vector2.one * pixelSize;
         }
